+++ src/rust/api.h
@@ -3,4 +3,4 @@ SEXP savvy_format_yaml_native__ffi(SEXP c_arg__value, SEXP c_arg__multi, SEXP c_
 SEXP savvy_init_yaml12__ffi(DllInfo* c_arg___dll_info);
 SEXP savvy_parse_yaml_native__ffi(SEXP c_arg__text, SEXP c_arg__multi, SEXP c_arg__simplify, SEXP c_arg__handlers);
 SEXP savvy_read_yaml_native__ffi(SEXP c_arg__path, SEXP c_arg__multi, SEXP c_arg__simplify, SEXP c_arg__handlers);
-SEXP savvy_write_yaml_native__ffi(SEXP c_arg__value, SEXP c_arg__path, SEXP c_arg__multi, SEXP c_arg__width);
+SEXP savvy_write_yaml_native__ffi(SEXP c_arg__value, SEXP c_arg__path, SEXP c_arg__multi, SEXP c_arg__width, SEXP c_arg__append);